Reduce Node sizes by shrinking "height" from Int32 to Byte

+++ src/System.Collections.Immutable/src/System/Collections/Immutable/ImmutableList`1.cs
@@ -1792,7 +1792,7 @@ namespace System.Collections.Immutable
             /// <summary>
             /// The depth of the tree beneath this node.
             /// </summary>
-            private int height;
+            private byte height; // AVL tree max height <= ~1.44 * log2(maxNodes + 2)
 
             /// <summary>
             /// The number of elements contained by this subtree starting at this node.
@@ -1843,7 +1843,7 @@ namespace System.Collections.Immutable
                 this.key = key;
                 this.left = left;
                 this.right = right;
-                this.height = 1 + Math.Max(left.height, right.height);
+                this.height = (byte)(1 + Math.Max(left.height, right.height));
                 this.count = 1 + left.count + right.count;
                 this.frozen = frozen;
             }
@@ -3215,7 +3215,7 @@ namespace System.Collections.Immutable
                         this.right = right;
                     }
 
-                    this.height = 1 + Math.Max(this.left.height, this.right.height);
+                    this.height = (byte)(1 + Math.Max(this.left.height, this.right.height));
                     this.count = 1 + this.left.count + this.right.count;
                     return this;
                 }
